Image Rotation Fix
[Problem]
On some Android devices, when the device rotation is locked,
the image is rotated 90 degrees.

[Solution]
Checked if the Exif:Orientation is 6 (rotated 90 degrees). If yes,
used Expo:ImageManipulation library to rotate the image -90
degrees, only on Android devices.

[Note]
This fixes the rotation of the image taken in portrait mode when
the device rotation is locked, but, as a consequence it also applies
the rotation to images taken in landscape mode when the device
rotation is locked. Unfortunately, there is no way to differentiate
between these two scenarios.

[Test]
Verified that the images taken in portrait mode when the device
rotation is locked shows up correctly. Also, verified that there
are no regressions while taking images when device rotation is
unlocked.